+++ vantage-sql/src/postgres/vista/factory.rs
@@ -64,6 +64,7 @@ impl PostgresVistaFactory {
             VistaCapabilities {
                 can_count: true,
                 can_order: true,
+                can_filter_operators: true,
                 can_insert: !read_only,
                 can_update: !read_only,
                 can_delete: !read_only,